Quick fix to invrep, added trie

+++ declaraciones/declaraciones.h
@@ -16,7 +16,7 @@ typedef vector<vii> wgraph;
 
 #define rep(i,n) for(size_t i = 0; i < (size_t)n; i++)
 #define repx(i,a,b) for(size_t i = a; i < (size_t)b; i++)
-#define invrep(i,a,b) for(size_t i = b; i-- > (size_t)a)
+#define invrep(i,a,b) for(int i = b; i-- > (int)a)
 
 #define pb push_back
 #define eb emplace_back
